actions: Split out zerver.actions.users.
