+++ zerver/management/commands/delete_user.py
@@ -3,7 +3,7 @@ from typing import Any
 
 from django.core.management.base import CommandError
 
-from zerver.lib.actions import do_delete_user
+from zerver.actions.users import do_delete_user
 from zerver.lib.management import ZulipBaseCommand
 from zerver.lib.users import get_active_bots_owned_by_user
