+++ zerver/tests/test_email_notifications.py
@@ -15,7 +15,8 @@ from django.test import override_settings
 from django.utils.timezone import now as timezone_now
 from django_auth_ldap.config import LDAPSearch
 
-from zerver.lib.actions import do_change_user_role, do_change_user_setting
+from zerver.actions.users import do_change_user_role
+from zerver.lib.actions import do_change_user_setting
 from zerver.lib.email_notifications import (
     enqueue_welcome_emails,
     fix_emojis,
