+++ zerver/tests/test_import_export.py
@@ -21,13 +21,13 @@ from zerver.actions.realm_icon import do_change_icon_source
 from zerver.actions.realm_logo import do_change_logo_source
 from zerver.actions.user_activity import do_update_user_activity, do_update_user_activity_interval
 from zerver.actions.user_topics import do_mute_topic
+from zerver.actions.users import do_deactivate_user
 from zerver.lib import upload
 from zerver.lib.actions import (
     check_add_reaction,
     do_add_reaction,
     do_change_realm_plan_type,
     do_create_user,
-    do_deactivate_user,
     do_mute_user,
 )
 from zerver.lib.avatar_hash import user_avatar_path
