+++ zerver/tests/test_message_send.py
@@ -10,16 +10,15 @@ from django.http import HttpResponse
 from django.test import override_settings
 from django.utils.timezone import now as timezone_now
 
+from zerver.actions.users import do_change_can_forge_sender, do_deactivate_user
 from zerver.lib.actions import (
     build_message_send_dict,
     check_message,
     check_send_stream_message,
     do_add_realm_domain,
-    do_change_can_forge_sender,
     do_change_stream_post_policy,
     do_create_realm,
     do_create_user,
-    do_deactivate_user,
     do_send_messages,
     do_set_realm_property,
     extract_private_recipients,
