+++ zerver/tests/test_signup.py
@@ -35,19 +35,17 @@ from zerver.actions.invites import (
     do_get_invites_controlled_by_user,
     do_invite_users,
 )
+from zerver.actions.users import change_user_is_active, do_change_user_role, do_deactivate_user
 from zerver.context_processors import common_context
 from zerver.decorator import do_two_factor_login
 from zerver.forms import HomepageForm, check_subdomain_available
 from zerver.lib.actions import (
     add_new_user_history,
-    change_user_is_active,
     do_change_full_name,
     do_change_realm_subdomain,
-    do_change_user_role,
     do_create_realm,
     do_create_user,
     do_deactivate_realm,
-    do_deactivate_user,
     do_set_realm_property,
     do_set_realm_user_default_setting,
     process_new_human_user,
