+++ zerver/views/development/email_log.py
@@ -9,11 +9,8 @@ from django.shortcuts import redirect, render
 from django.views.decorators.http import require_safe
 
 from confirmation.models import Confirmation, confirmation_url
-from zerver.lib.actions import (
-    change_user_is_active,
-    do_change_user_delivery_email,
-    do_send_realm_reactivation_email,
-)
+from zerver.actions.users import change_user_is_active
+from zerver.lib.actions import do_change_user_delivery_email, do_send_realm_reactivation_email
 from zerver.lib.email_notifications import enqueue_welcome_emails
 from zerver.lib.response import json_success
 from zerver.models import Realm, get_realm, get_realm_stream, get_user_by_delivery_email
