+++ zerver/views/users.py
@@ -10,6 +10,12 @@ from zerver.actions.custom_profile_fields import (
     check_remove_custom_profile_field_value,
     do_update_user_custom_profile_data_if_changed,
 )
+from zerver.actions.users import (
+    do_change_user_role,
+    do_deactivate_user,
+    do_update_bot_config_data,
+    do_update_outgoing_webhook_service,
+)
 from zerver.context_processors import get_valid_realm_from_request
 from zerver.decorator import require_member_or_admin, require_realm_admin
 from zerver.forms import PASSWORD_TOO_WEAK_ERROR, CreateUserForm
@@ -21,13 +27,9 @@ from zerver.lib.actions import (
     do_change_default_all_public_streams,
     do_change_default_events_register_stream,
     do_change_default_sending_stream,
-    do_change_user_role,
     do_create_user,
-    do_deactivate_user,
     do_reactivate_user,
     do_regenerate_api_key,
-    do_update_bot_config_data,
-    do_update_outgoing_webhook_service,
     notify_created_bot,
 )
 from zerver.lib.avatar import avatar_url, get_gravatar_url
